make UDC.NewSession 1 screen + add edit logic with bug
