+++ db/schemas/userSchema.js
@@ -3,6 +3,7 @@
  */
 const mongoose = require('../../server/node_modules/mongoose');
 const bcrypt = require('../../server/node_modules/bcrypt');
+import { UDCSchema } from './UDCSchema';
 
 export const UserSchema = new mongoose.Schema({
     email: {
@@ -14,6 +15,9 @@ export const UserSchema = new mongoose.Schema({
     password: {
         type: String,
         required: true,
+    },
+    sessions: {
+      UDC: [{...UDCSchema}]
     }
 });
