+++ mobile/constants/Api.js
@@ -4,25 +4,26 @@ HEROKU_API_URL = 'https://doggoplatform.herokuapp.com/api/';
 LOCAL_API_URL = 'http://localhost:3010/api/';
 
 // routes
-USERS_ROUTE = 'users';
-DATA_COLLECTION_ROUTE = 'data_collection';
-
+USERS_ROUTE = 'users/';
+SESSIONS_ROUTE = 'sessions';
 
 const formatAPILink = url => LOCAL_API_URL + url;
+const formatSessionsRoute = route => (
+  new Promise((res, rej) => {
+    getUserID()
+      .then((id) => (
+          res(formatAPILink(`${USERS_ROUTE}/${id}/${SESSIONS_ROUTE}/${route}`))
+      ))
+      .catch((err) => rej(err))
+  })
+)
+
 export default {
-    getUDCSession: formatAPILink(DATA_COLLECTION_ROUTE + '/get-UDC-session'),
-    getCurrentUDCSessions: formatAPILink(DATA_COLLECTION_ROUTE + '/get-current-UDC-sessions'),
-    getLoginApiURL: formatAPILink(USERS_ROUTE + '/login'),
-    getRegisterApiURL: formatAPILink(USERS_ROUTE + '/register'),
-    getLogoutApiURL: formatAPILink(USERS_ROUTE + '/logout'),
-    getProfileApiURL: formatAPILink(USERS_ROUTE + '/profile'),
-    getSaveUDCSessionURL: () => {
-        return new Promise((res, rej) => {
-            getUserID()
-                .then((id) => (
-                    res(formatAPILink(`${DATA_COLLECTION_ROUTE}/${id}/create-UDC-session`))
-                ))
-                .catch((err) => rej(err))
-        })
-    }
+  getUDCSession: formatAPILink(formatSessionsRoute('udc/get-session')),
+  getCurrentUDCSessions: formatSessionsRoute('udc/get-current-sessions'),
+  loginApiURL: formatAPILink(USERS_ROUTE + 'login'),
+  registerApiURL: formatAPILink(USERS_ROUTE + 'register'),
+  logoutApiURL: formatAPILink(USERS_ROUTE + 'logout'),
+  profileApiURL: formatAPILink(USERS_ROUTE + 'profile'),
+  saveUDCSessionURL: formatSessionsRoute('udc/create-new-session')
 }
