+++ mobile/redux/actions/index.actions.js
@@ -1,5 +1,5 @@
-export const GET_UDC_GENERAL_INITIAL_STATE = 'GET_UDC_GENERAL_INITIAL_STATE';
-export const GET_UDC_HIDES_INITIAL_STATE = 'GET_UDC_HIDES_INITIAL_STATE';
-export const SAVE_NEW_UDC_SESSION = 'SAVE_NEW_UDC_SESSION';
+// export const GET_UDC_NEW_SESSION_INITIAL_STATE = 'GET_UDC_NEW_SESSION_INITIAL_STATE';
+export const SAVE_UDC_SESSION = 'SAVE_UDC_SESSION';
+export const EDIT_NEW_UDC_SESSION = 'EDIT_NEW_UDC_SESSION';
 export const UPDATE_NEW_UDC_SESSION_HIDE = 'UPDATE_NEW_UDC_SESSION_HIDE_STATE';
